[DOS] Introducing "-vanilla” parameter
I want everyone to understand that this is not, and will not be complete
vanilla emulation. This is just disables some additional gameplay
features, to make game feeling more vanilla-like, so let’s just call
this option “moderate purism”. Besides, this also fits to main project’s
Philosophy, because I always want to give a choice to the player - want
he/she see&feel new features or not.

This parameter disables:
- Colored startup banners (they will be looking as vanilla 1.9 versions)
- All shadowed texts and menu elements
- No prompting for quick saving/loading
- Fix for fast door making two closing sounds
- Fix for fast doors reopening with wrong sound
- Playing sound effect when the doors are closed manually
- Fix for firing sound breaks by player’s “oof!”
- Lethal pellet of a point-blank SSG blast
- SFX of crushed corpse
- Randomization of dead marine colors
- Randomization of deaths and corpses flipping
- Notification of discovered secrets
- Total level times on intermission screen
- Swirling liquids
- INVUL sphere now affecting the sky
- Preventing IDDQD for dead player

This parameter will not do anything about:
- Static engine limits
- Technical engine improvements (wiggle and so on)
- New menu elements (colored sliders, 8 saving slots, new titles for
SFX/Save/Load)
- Golden eyes while grinning/rampage in god mode
- Status bar hysteresis fix

In any way, technically it’s a huge hack, rather than supreme feature.
But this is a right thing to do.

Finally, I would like to thank @bradharding for the Idea of making game
feeling more vanilla by exactly one parameter.