[DOS] Introducing -solo-coop parameter
@fabiangreffrath, @bradharding, @axdoomer, @jeffdoggett,
colleagues, take a look, maybe you’ll find this idea interesting or
useful somehow for your projects.

This will activate cooperative spawn factor in single player.
Nothing else, no respawns, no resurrections.

It is not the same as Chocolate Doom "-solo-net", because it does not
block player's ability to "load saved game in game", and also,
"-solo-net" implementation is not working in DOS version - I've getting
a heavy DOSBox lockup with sound mess.

Why is it necessary?

Exclusively for making some *optional* vanilla difference in gameplay by
vanilla's way. For example, now it is possible to meet Mastermind in
MAP12 and Cyberdemon in MAP13.

Some more ideas:

- I need to understand, do I need singleplayer condition here or not.
From one hand, having a coop monsters in vanilla demos is apocalypse.
From another, this game mode is not supposed to watch vanilla demos, and
I don't want to prohibit anyone to record demos in this mode (despite of
0% will be interested to make a demo records using my port).
- This mode also spawns coop weapons, so you can get BFG right in
Entryway. Maybe it is possible to make some flag like "!weapons", but
these weapons are placed in coop by design. And it's definitely better
to have a one command line parameter for this mode, rather than
something like "-solo-coop -solo-coop-no-weapons" and so on. In any way,
player always can get more challenge by setting on fast monsters and
higher skill levels.