Disabled Strife support
I'm not sure I will finish this translation. It was interesting to draw
alphabet, break head on menu positions, making other nice stuff, but
another side of the medal: it's a pure nightmare to work with Slade's
limitations of dialogue texts. Also, this game have a lots of speech
that must be not only translated, but properly adapted to Russian
language, and doing this without good knowledge of game’s universe is a
bad idea. IDK, maybe I will finish this translation one day, but
probably this day will never come.