Added animated candle
Seems to be safe for mod compatibility, but checking is not very proper yet.

For example, candle will not be animated with "-file" usage, and for some reason, this condition is not working:

if (W_CheckNumForName("SS_START") >= 0 && W_CheckNumForName("CANDA0") >= 0)

Need to break my head more, otherwise this feature will be available only in clean, totally unmodified game.