Implemented brightmaps. Hell yeah!
Well, this is it. I'm using different COLORMAPs for different sprites and walls.

There is nothing extraordinary here, in fact, all additional COLORMAP's calculation is based on standard formulas, which seems to be safe. At least, everything working properly and there is no performance drops.

I'm deeply thankful to @bradharding for his research of brightmap tables and colors - my colormaps are based on them. :)

TODO:
- Make brightmaps optional through setup executable. Too tired to do it now.
- Apply "redonly" brightmap to the hud's BFG. At least for some states.
- Port this code to DOS version. This could be a bomb.