Fake contrast: make optional, disabled by default
Unbelievable. I was thinking that the game will looks somewhat “flat” without fake contrast, but in fact it looks even better and more realistic. Probably that’s because of combination of Crispy’s awesome middle-resolution and my improved palette. Original diminished lightning is also making a good things in visual.

As for DOS version, better to leave fake contrast “as is”, since pixels there are always in 320x200 and I’m using my previous implementation of palette improvement, which is less darken for better visual in DOSbox and probably in some older non too bright monitors.