Press Beta: fixed very rare vanilla bug on map №2
One of black rooms (with demons) has been made slightly wider, in the section with pentagram on floor. This will allow to remaining demons to don't block each other and teleport normally.

In theory, vanilla E3M5 is also affected, but this seems to be code, not mapping related bug.

Step №10 in #50.