`geartics` revisited
Now geartics are decreased only while torque effect is active, i.e. not while simple falling or sliding because of momentum received.

Tics are also increased to 15 seconds. This is probably a reasonable middle between "infinite sliding" and "hanging partway". Some casual playtesting is still needed to see is it good or not good enough.

Personal thanks to @bradharding for pointing my thoughts to right direction. :)