Always show full list of bound keys in menu
It appears that reducing number of shown keys to one in non-wide resolutions is a bad idea. Most of the key names are short, and it's annoying what you can't see all binds.
For long names I think it's better to let them go beyond the screen for now in exchange for usability in other cases and deal with them later. Maybe by implementing the scrolling text feature.