Windows: Allow text selection and copying in a console window
This has a side effect: when something is selected in the console window, OS will halt the thread attempting to write to the console until the selection is removed.

In real-case scenarios this side effect is quite rare as when you select something, you probably want to copy it and after copying the selection is removed automatically.