Windows: Split `.exe` to `.dll` and `.exe`/`.com`
Inspired by JDK, and it's `java.exe` and `javaw.exe`.

The majority of the port's code compiles to `.dll`. The `.exe` and `.com` executables link this DLL and delegate execution to it.
This eliminates creation of an extra process and hopefully will shut up antivirus false positives.
The `.exe` is still compiled as `/SUBSYSTEM:WINDOWS` and `.com` is still compiled as `/SUBSYSTEM:CONSOLE` to provide "clean" GUI and proper CLI in most use cases.