Replace `libpng` and `zlib` with `miniz`
`libpng` and `zlib` were the last libraries that depend on `libgcc_s_dw2-1.dll` which breaks Make, CLion and other tools by simply existing in the working directory. And now we can remove them all from the distribution.
Also, `libpng` causes crushes on clang builds when user takes screenshot because of CRT conflict.

Usually `miniz` will be compiled into Inter-doom executable, but if it exists on the system as a shared library, we will use the shared library.