Remove sources for `setup.exe`
It doesn't understand new config structure, doesn't know about many new config entries because it wasn't updated for a long time. So it is unusable even if compiled.
Most of it's features already implemented in game menu except the launcher functionality that I will implement differently.

If code for `setup.exe` is needed for reference, it can always be viewed through the git history.