Change versioning of dev builds
Because Doom, Heretic and Hexen have their own release cycles and version numbers that may or may not overlap, we can't use one of the releases as a base for dev build version. The "6.x.x" works only until the major version grows beyond 6 and then must be replaced by the new major version. This may be sufficient for GitHub actions, but for the `-git` AUR packages, the version should be determined by a function. Many AUR packages use a combination of number of commits on the master branch and short hash of the last commit, so I think the same approach will be good for us.

Due to changes made to `revision_check`, there is no more need for building and packaging only one game at a time.

fixup! Change versioning of dev builds