Hexen: Restore original transparency algorithm for fullscreen inventory bar and armor icons on automap
There are two algorithms:
1) We call "dest algorithm"
```
*dest = tinttable[((*dest) << 8) + *source];
```
2) We call "source algorithm"
```
*dest = tinttable[*dest + ((*source)<<8)];
```

In the original Heretic, `V_DrawTLPatch` was called `V_DrawFuzzPatch` and used the "dest algorithm".
In the original Hexen, `V_DrawTLPatch` was called `V_DrawFuzzPatch` and used the "source algorithm",
and `V_DrawAltTLPatch` was coled `V_DrawAltFuzzPatch` and used the "dest algorithm".

So to preserve original behavior, we need to leave `V_DrawTLPatch` with "dest algorithm",
implement `V_DrawAltTLPatch` with "source algorithm" and
swap invocations of `V_DrawTLPatch` and `V_DrawAltTLPatch` in Hexen.