Fix config corruption if Windows username contains non-English characters
The "player_name" config parameter is initialized with the value of "USER" or "USERNAME" environment variables.
On Windows, it is in OEM encoding, so it needs to be converted to UTF-8.
However, this conversion was done every time, not just for the first initialization, resulting in an ever-growing string when it contained non-English characters.