Change how CLI arguments are processed
Arguments may start with lazy parameters, that may be followed by the regular parameters (and/or response file).
Lazy parameters are the paths to wads, deh/bex patches, demos that will be heuristically categorized and loaded.
Regular parameters (and response file) start with keywords prefixed with '-' (or '@' for a response file) and expected to be well-formed.